refactor(camera)!: move views to separate package
