+++ lib/src/main/java/io/github/nonmilk/coffee/grinder/camera/PerspectiveCamera.java
@@ -1,6 +1,7 @@
 package io.github.nonmilk.coffee.grinder.camera;
 
 import io.github.alphameo.linear_algebra.mat.Matrix4;
+import io.github.nonmilk.coffee.grinder.camera.view.PerspectiveView;
 
 /**
  * Represents a perspective projection camera.
